% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_func.R
\name{.F_func}
\alias{.F_func}
\title{Auxiliary function used by .ZI_projection}
\usage{
.F_func(rho, x, y, s, z, A, b, cl, p, B, xe, ka)
}
\arguments{
\item{rho}{g}

\item{x}{g}

\item{y}{g}

\item{s}{g}

\item{z}{g}

\item{A}{g}

\item{b}{g}

\item{cl}{g}

\item{p}{g}

\item{B}{g}

\item{xe}{g}

\item{ka}{g}
}
\description{
Auxiliary function used by .ZI_projection
}
\keyword{internal}
