% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_diag.R
\name{.make_diag}
\alias{.make_diag}
\title{Makes a diagonal matrix with values from x}
\usage{
.make_diag(x)
}
\arguments{
\item{x}{vector}
}
\value{
diagonal matrix y
}
\description{
Function make_diag() takes a vector x and returns a matrix with diagonal consists the values of x.
}
\keyword{internal}
