% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RISK_post.R
\name{.RISK_post}
\alias{.RISK_post}
\title{Computes the empirical Conditional Value-at-Risk, Value-at-Risk and Mean Absolute Deviation for losses with given probabilities}
\usage{
.RISK_post(returns, prob, alpha)
}
\arguments{
\item{returns}{vector of losses}

\item{prob}{vector of probability of losses}

\item{alpha}{confidence level for CVaR and VaR}
}
\value{
list of values
}
\description{
Computes the empirical Conditional Value-at-Risk, Value-at-Risk and Mean Absolute Deviation for losses with given probabilities
}
\keyword{internal}
