% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZI_projection.R
\name{.ZI_projection}
\alias{.ZI_projection}
\title{Computes \cr
the solution of the linear program
\eqn{\min c^{T} x}
 for \eqn{Ax = b}, \eqn{x \ge 0}}
\usage{
.ZI_projection(clin, Amat, bmat, xhat, B, maxiter, tol, k)
}
\arguments{
\item{clin}{g}

\item{Amat}{g}

\item{bmat}{g}

\item{xhat}{g}

\item{B}{g}

\item{maxiter}{g}

\item{tol}{g}

\item{k}{g}
}
\value{
list
}
\description{
which fulfills the condition \eqn{|B(x - xhat)|^2 \to \min}
}
\keyword{internal}
